% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- a numeric vector}
}
\value{
Value that occurs most frequently in given numeric vector x. NA elements are ignored, if given vector x contains only NA elements then NA value will be returned.
}
\description{
This function determines the most frequently occurring value in an integer vector (mode).
}

