% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{simplify2array}
\usage{
simplify2array(x)
}
\arguments{
\item{x}{- a list}
}
\value{
A matrix, numeric vector or a list. If it is feasible to simplify (all elements of given x list are numeric vectors of the same length) then matrix or numeric vector is returned. Otherwise function returns unchanged given list x.
}
\description{
This function tries to convert list to matrix or numeric vector.
}

